// iMusic/ClientApp/src/app/core/swagger/model/songDTO.ts
/**
 * iMusic DTOs
 * iMusic
 *
 * OpenAPI spec version: 1.0.0
 * Contact: TODO: add iMusic email
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ArtistDTO } from './artistDTO';

export interface SongDTO { 
    songId?: number;
    name?: string;
    length?: string;
    price?: number;
    image?: string;
    artists?: Array<ArtistDTO>;
}